requirements.txt
pandas
numpy
google-api-python-client~=2.149.0
httplib2~=0.22.0
oauth2client~=4.1.3
google-auth-oauthlib~=1.2.1
protobuf~=5.28.2
google
openpyxl
aiogram
python-dotenv